Create  GetElementPosition - Core Java
Given a method with an array of strings and one string variable as input. Write code to sort the given array in reverse alphabetical order and return the postion of the given string in the array. Include a class UserMainCode with a static method getElementPosition which accepts an array of strings and a string variable as input. The return type of the output is an integer which is the position of given string value from the array. Create a Main class which gets string array and a string variable as an input and call the static method getElementPosition present in the UserMainCode.