Create differenceMax&MinArray
Given a method taking an int array having size more than or equal to 1 as input. Write code to return the difference between the largest and smallest elements in the array. If there is only one element in the array return the same element as output. Include a class UserMainCode with a static method getBigDiff which accepts a integer array as input. The return type of the output is an integer which is the difference between the largest and smallest elements in the array. Create a Main class which gets integer array as an input and call the static method getBigDiff present in the UserMainCode.