Create ObjectasParameters
As you can see, the equalTo( ) method inside Test compares two objects for equality
and returns the result. That is, it compares the invoking object with the one that it is
passed. If they contain the same values, then the method returns true. Otherwise, it returns
false. Notice that the parameter o in equalTo( ) specifies Test as its type. Although Test is a
class type created by the program, it is used in just the same way as Java’s built-in types.