Create Digits
Write a program to read a non-negative integer n, that returns the count of the occurances of 7 as digit. Include a class UserMainCode with a static method countSeven which accepts the integer value. The return type is integer which is the count value. Create a Class Main which would be used to accept the string and call the static method present in UserMainCode.