Create reverseANDtruncate
Given a string, startIndex and length, write a program to extract the substring from right to left. Assume the last character has index 0. Include a class UserMainCode with a static method ?reverseSubstring? that accepts 3 arguments and returns a string. The 1st argument corresponds to the string,the second argument corresponds to the startIndex and the third argument corresponds to the length