Create Triplets
Given an integer array, Write a program to find if the array has any triplets. A triplet is a value if it appears 3 consecutive times in the array. Include a class UserMainCode with a static method checkTripplets which accepts an integer array. The return type is boolean stating whether its a triplet or not. Create a Class Main which would be used to accept the input arrayand call the static method present in UserMainCode.