Create DayOfWeekafterAyear
Given a date string in dd/mm/yyyy format, write a program to calculate the day which falls on the same date next year. Print the output in small case.The days are sunday, monday, tuesday, wednesday, thursday, friday and saturday.
Include a class UserMainCode with a static method ?nextYearDay? that accepts a String and returns a String.