Create RemoveIntFromString
Write a program to input a String array. The input may contain digits and alphabets (“de5g4G7R”). Extract odd digits from each string and find the sum and print the output. For example, if the string is ""AKj375A"" then take 3+7+5=15 and not as 375 as digit. Include a class UserMainCode with a static method oddDigitSum which accepts a string array and the size of the array. The return type (Integer) should return the sum. Create a Class Main which would be used to accept Input Strings and call the static method present in UserMainCode. Assume maximum length of array is 20.